ast: move the big macros to their own module
